+++ app/(survey)/(survey-funnel)/_steps/enter-life-satisfaction-step.tsx
