+++ app/(survey)/(survey-funnel)/_steps/start-step.tsx
