+++ constants/routes.ts
@@ -1,5 +1,6 @@
 const ROUTES = {
   ROOT: '/',
+  DONE: '/done',
 } as const;
 
 export { ROUTES };
